deleted UUID, not connected with PostgreSql. fixed column dependencies

+++ src/main/java/codezilla/handynestproject/repository/WorkingTimeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkingTimeRepository extends JpaRepository<WorkingTime, Integer> {
+public interface WorkingTimeRepository extends JpaRepository<WorkingTime, Long> {
 }
